Extract modules list table to be used both as a page or widget
