+++ src/presentation/widget/pages/Root.tsx
@@ -1,5 +1,5 @@
 import qs from "qs";
-import React from "react";
+import React, { ReactNode, useEffect, useState } from "react";
 
 export function useWidget() {
     const { dashboardItemId = "DEVELOPMENT", userOrgUnit } = qs.parse(window.location.search);
@@ -7,12 +7,32 @@ export function useWidget() {
     return {
         dashboardItemId,
         userOrgUnits: userOrgUnit?.split(",") ?? [],
+        widget: "modules-list",
     };
 }
 
+const loadWidget = async (widget: string) => {
+    switch (widget) {
+        case "modules-list":
+            const { ModulesListWidget } = await import("./modules-list-widget/ModulesListWidget");
+            return ModulesListWidget;
+        default:
+            return () => {
+                const { dashboardItemId } = useWidget();
+                return <p>{`Hello World, I'm dashboard item ${dashboardItemId}!`}</p>;
+            };
+    }
+};
+
 function Root() {
-    const { dashboardItemId } = useWidget();
-    return <p>{`Hello World, I'm dashboard item ${dashboardItemId}!`}</p>;
+    const { widget } = useWidget();
+    const [Component, setComponent] = useState<ReactNode>(null);
+    
+    useEffect(() => {
+        loadWidget(widget).then(setComponent);
+    }, [widget]);
+
+    return Component;
 }
 
 export default Root;
