+++ src/presentation/widget/pages/modules-list-widget/ModulesListWidget.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { ModulesListTable } from "../../../common/components/modules-list-table/ModulesListTable";
+
+export const ModulesListWidget: React.FC = () => {
+    return (
+        <React.Fragment>
+            <ModulesListTable />
+        </React.Fragment>
+    );
+};
